use HEAD instead of master branch
The master branch could go away in the future. Especially as main becomes the
standard default branch for repositories going forward. This allows the scripts
to still reference the tip of the default branch without relying on the name of
the branch. HEAD is just a ref to the tip of the working branch. On GitHub, the
working branch is the repository's default branch. e.g. spacemacs' default
branch is develop instead of master.